Mostly done

+++ core/src/main/resources/app/scripts/services.js
@@ -1,7 +1,41 @@
 'use strict';
 
+function mkDebt(service, $http, userSrv, errSrv) {
+  service.payDebt = function (debt) {
+    if (debt.paid) {
+      errSrv.emit("Debt has already been marked as paid.");
+    } else if (debt.debtor.id != userSrv.user.id) {
+      errSrv.emit("You cannot repay someone else's debt.");
+    } else {
+      $http.post("api/debts/pay/" + debt.id)
+      .then(function (response) {
+        debt.paid = true;
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+    }
+  };
+
+  service.confirmDebt = function (debt) {
+    if (!debt.paid) {
+      errSrv.emit("Cannot confirm an unpaid debt.");
+    } else if (debt.confirmed) {
+      errSrv.emit("Debt has already been confirmed.");
+    } else if (debt.creditor.id != userSrv.user.id) {
+      errSrv.emit("You cannot confirm a debt not owed to you.");
+    } else {
+      $http.post("api/debts/confirm/" + debt.id)
+      .then(function (response) {
+        debt.confirmed = true;
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+    }
+  };
+};
+
 angular.module('moneyApp')
-  .factory('userSrv', function($http, $q, errSrv) {
+  .factory('userSrv', function($http, $q, $rootScope, errSrv) {
     var service = { user: {}, activeUsers: [] };
 
     var defaultUser = {
@@ -23,7 +57,7 @@ angular.module('moneyApp')
       }
     };
 
-    service.populateActiveUsers = function(data) {
+    service.populateActiveUsers = function (data) {
       console.log("active users" + data.length);
       for (var i = 0; i < data.length; i++) {
         service.activeUsers.push(data[i]);
@@ -31,11 +65,12 @@ angular.module('moneyApp')
       }
     };
 
-    service.getActiveUsers = function () {
+    service.getActiveUsers = function (callback) {
       $http.post('api/activeUsers')
       .then(function (response) {
         service.resetActiveUsers();
         service.populateActiveUsers(response.data);
+        callback();
       }, function (err) {
         errSrv.emit(err.data);
       });
@@ -58,7 +93,17 @@ angular.module('moneyApp')
       $.extend(service.user, defaultUser);
     };
 
-    service.getActiveUsers();
+    service.checkActiveLogin = function (success, failure) {
+      $http.get('api/whoami')
+      .then(function (response) {
+        service.populateUserData(response.data);
+        success();
+      }, function (err) {
+        console.log("No login detected: " + err.data);
+        service.reset();
+        failure();
+      });
+    }
 
     return service;
   })
@@ -104,33 +149,39 @@ angular.module('moneyApp')
 
     return service;
   })
-  .factory('redirectSrv', function($location) {
+  .factory('redirectSrv', function ($location) {
     var service = {};
     service.redirectUrl = null;
 
-    service.redirect = function() {
+    service.redirect = function () {
       if (service.redirectUrl) {
         $location.path(service.redirectUrl);
         service.redirectUrl = null;
       }
     }
 
-    service.setup = function(url) {
+    service.setup = function (url) {
       service.redirectUrl = url;
+    };
+
+    service.orElse = function (url) {
+      if (!service.redirectUrl) {
+        service.redirectUrl = url;
+      }
     };
 
     return service;
   })
-  .factory('errSrv', function($rootScope) {
+  .factory('errSrv', function ($rootScope) {
     var service = {};
 
     service.errors = [];
 
-    service.emit = function(err) {
+    service.emit = function (err) {
       service.errors.push(err);
     }
 
-    service.reset = function() {
+    service.reset = function () {
       while(service.errors.length > 0) {
           service.errors.pop();
       }
@@ -138,31 +189,35 @@ angular.module('moneyApp')
 
     return service;
   })
-  .factory('expenseSrv', function($http, errSrv, userSrv, redirectSrv) {
+  .factory('expenseSrv', function ($http, errSrv, userSrv, redirectSrv) {
     var service = {};
-
-    service.activeUsers = userSrv.activeUsers;
 
     service.state = { addOrEdit: "Add", errLoading: false };
 
     service.expense = { }
 
-    service.defaultExpense = {
-      id: null,
-      owner: userSrv.user,
-      date: null,
-      amount: 0.0,
-      description: "description",
-      comment: "comment",
-      recurring: false,
-      frequencyNum: 1,
-      frequencyUnit: 0,
-      beneficiaries: service.activeUsers.map(function (au) { return $.extend({ selected: true }, au); }),
-      selected: false
+    service.defaultExpense = function () {
+      var res = {
+        id: null,
+        owner: userSrv.user,
+        date: null,
+        amount: 0.0,
+        description: "description",
+        comment: "comment",
+        recurring: false,
+        frequencyNum: 1,
+        frequencyUnit: 0,
+        beneficiaries: userSrv.activeUsers.map(function (au) { return $.extend({ selected: true }, au); }),
+        selected: false
+      };
+
+      console.log("HOORAH: " + JSON.stringify(userSrv.activeUsers));
+
+      return res;
     };
 
     service.reset = function (rec) {
-      $.extend(service.expense, service.defaultExpense);
+      $.extend(service.expense, service.defaultExpense());
       service.expense.recurring = rec;
       service.state.errLoading = false;
       service.state.addOrEdit = "Add";
@@ -198,10 +253,14 @@ angular.module('moneyApp')
           service.reset(false);
           service.state.errLoading = true;
           errSrv.emit("You do not have permissions to edit this expense.");
+        } else if (response.data.committed) {
+          service.reset(false);
+          service.state.errLoading = true;
+          errSrv.emit("You cannot edit a committed expense.");
         } else {
           service.state.errLoading = false;
           $.extend(service.expense, response.data);
-          service.expense.beneficiaries = service.activeUsers.map(function (au) {
+          service.expense.beneficiaries = userSrv.activeUsers.map(function (au) {
             var isBen = service.expense.beneficiaries.filter(function (ben) { return au.id == ben.id; }).length == 1;
             return $.extend({ selected: isBen }, au);
           });
@@ -319,7 +378,11 @@ angular.module('moneyApp')
         var msg = "Are you sure you want to delete the selected expenses?";
         var onSuccess = function () {
           selectedExpenses.forEach(function (expense) {
-            if (userSrv.user.admin || userSrv.user.id == expense.owner.id) {
+            if (!userSrv.user.admin && userSrv.user.id != expense.owner.id) {
+              errSrv.emit("You do not have the permissions to delete this expense.");
+            } else if (expense.committed) {
+              errSrv.emit("You do not delete a committed expense.");
+            } else {
               $http.delete('api/expense/' + expense.id).then(function (response) {
                 service.expenses.forEach(function (value, index) {
                   if (value.id == expense.id) {
@@ -329,8 +392,6 @@ angular.module('moneyApp')
               }, function (err) {
                 errSrv.emit(err.data);
               });
-            } else {
-              errSrv.emit("You do not have the permissions to delete this expense.");
             }
           });
         }
@@ -362,7 +423,7 @@ angular.module('moneyApp')
     service.cancel = function () { service.hide(); }
     service.submit = function () { service.hide(); }
 
-    service.mkModal = function(msg, action) {
+    service.mkModal = function (msg, action) {
       service.msg = msg;
       service.submit = function () {
         service.hide();
@@ -373,23 +434,171 @@ angular.module('moneyApp')
     };
 
     return service;
-  }).factory('reportSrv', function ($http) {
+  }).factory('reportSrv', function ($http, userSrv, errSrv) {
     var service = {};
 
+    var emptyReport = {
+      id: null,
+      date: null,
+      description: null
+    };
+
     service.report = {};
+    service.expenses = [];
+    service.debts = [];
+
+    service.reset = function () {
+      while(service.expenses.length > 0) {
+        service.expenses.pop();
+      }
+
+      while(service.debts.length > 0) {
+        service.debts.pop();
+      }
+
+      $.extend(service.report, emptyReport);
+    };
+
+    service.populateReport = function (data) {
+      $.extend(service.report, data);
+    };
+
+    service.populateExpenses = function (data) {
+      // TODO: Dedup
+      data.forEach(function (expense) {
+        expense.beneficiaries = expense.beneficiaries.map(function (user) { return user.name; }).sort().join(", ");
+        service.expenses.push(expense);
+      });
+    };
+
+    service.populateDebts = function (data) {
+      data.forEach(function (debt) {
+        service.debts.push(debt);
+      });
+    };
+
+    service.getReport = function (id) {
+      service.reset();
+
+      $http.get('api/report/' + id)
+      .then(function (response) {
+        service.populateReport(response.data);
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+
+      $http.get('api/expense/report/' + id)
+      .then(function (response) {
+        service.populateExpenses(response.data);
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+
+      $http.get('api/debts/report/' + id)
+      .then(function (response) {
+        service.populateDebts(response.data);
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+    };
+
+    mkDebt(service, $http, userSrv, errSrv);
 
     return service;
-  }).factory('reportsSrv', function ($http) {
+  }).factory('reportsSrv', function ($http, $location, $filter, errSrv, modalSrv) {
     var service = {};
 
     service.reports = [];
 
+    service.reset = function () {
+      while(service.reports.length > 0) {
+        service.reports.pop();
+      }
+    };
+
+    service.populateReports = function (data) {
+      data.forEach(function (report) {
+        service.reports.push(report);
+      });
+    };
+
+    service.getReports = function () {
+      service.reset();
+
+      $http.get('api/reports')
+      .then(function (response) {
+        service.populateReports(response.data);
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+    };
+
+    service.generate = function (fromDateRaw, toDateRaw) {
+      if (fromDateRaw && toDateRaw) {
+        var fromDate = $filter('date')(fromDateRaw, "yyyy-MM-dd");
+        var toDate = $filter('date')(toDateRaw, "yyyy-MM-dd");
+        var msg = "Are you sure you want to commit all expenses from " + fromDate + " to " + toDate + "? This operation cannot be undone.";
+        var onSuccess = function () {
+          $http.post("api/report/generate/" + fromDate + "/" + toDate)
+          .then(function (response) {
+            // TODO: redirect to the newly-created report?
+            $location.path("/reports/list");
+          }, function (err) {
+            errSrv.emit(err.data);
+          });
+        };
+
+        modalSrv.mkModal(msg, onSuccess);
+      }
+    };
+
     return service;
-  }).factory('debtsSrv', function ($http) {
+  }).factory('debtsSrv', function ($http, userSrv, errSrv) {
     var service = {};
 
     service.myDebts = [];
     service.otherDebts = [];
+
+    service.reset = function () {
+      while(service.myDebts.length > 0) {
+        service.myDebts.pop();
+      }
+      while(service.otherDebts.length > 0) {
+        service.otherDebts.pop();
+      }
+    };
+
+    service.populateMyDebts = function (data) {
+      data.forEach(function (debt) {
+        service.myDebts.push(debt);
+      });
+    };
+
+    service.populateOtherDebts = function (data) {
+      data.forEach(function (debt) {
+        service.otherDebts.push(debt);
+      });
+    };
+
+    service.getDebts = function () {
+      service.reset();
+
+      $http.get('api/debts/debtor/' + userSrv.user.id)
+      .then(function (response) {
+        service.populateMyDebts(response.data);
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+
+      $http.get('api/debts/creditor/' + userSrv.user.id)
+      .then(function (response) {
+        service.populateOtherDebts(response.data);
+      }, function (err) {
+        errSrv.emit(err.data);
+      });
+    };
+
+    mkDebt(service, $http, userSrv, errSrv);
 
     return service;
   });
